# specs/001-zero-training-mvp/contracts/api-openapi.yaml
openapi: 3.1.0
info:
  title: LiverRa v1 MVP HTTP API
  version: 1.0.0
  description: |
    Public HTTP API for the LiverRa zero-training liver AI pipeline platform.

    **Auth**: OIDC Bearer token from AWS Cognito (research A.1) with `custom:tenant_id` claim.
    **RBAC**: Every protected op declares `x-required-permission`. Step-up auth (≤5 min freshness) is marked `x-step-up: true`.
    **Errors**: RFC 7807 Problem Details. `type`, `title`, `status`, `detail`, plus LiverRa extensions `tenant_id`, `audit_event_id`, `claim_key` (never PHI).
    **Audit**: Every state-changing op emits a FHIR AuditEvent (chain-of-hashes research A.3).
servers:
  - url: https://app.liverra.ai/api/v1
    description: Production
  - url: http://localhost:7050/api/v1
    description: Local dev (FastAPI ml-inference)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string, description: "MUST NOT contain PHI" }
        tenant_id: { type: string, format: uuid }
        audit_event_id: { type: string }
...
    Lesion:
      type: object
      properties:
        id: { type: string, format: uuid }
        couinaud_location: { type: string }
        longest_diameter_mm: { type: number }
        volume_ml: { type: number }
        discovery_source: { type: string, enum: [ai_detected, reviewer_prompted] }
        classification:
          type: object
          properties:
            suggested_class: { type: string, enum: [hcc, icc, metastasis, fnh, hemangioma, cyst, abstained] }
            confidence_vector:
              type: object
              additionalProperties: { type: number }
            abstention_threshold_used: { type: number }
            temperature_applied: { type: number }
            model_version: { type: string }
            reviewer_override_class: { type: string, nullable: true }
    FLRCalculation:
      type: object
      properties:
        id: { type: string, format: uuid }
        plane_normal: { type: array, items: { type: number }, minItems: 3, maxItems: 3 }
        plane_offset_mm: { type: number }
        resected_volume_ml: { type: number }
        remnant_volume_ml: { type: number }
        remnant_pct_functional: { type: number }
        author: { type: string, enum: [ai_default, surgeon_edited] }
    SurgeonReview:
      type: object
      properties:
        id: { type: string, format: uuid }
        analysis_id: { type: string, format: uuid }
        reviewer_user_id: { type: string, format: uuid }
        seat_held_until: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finalized_at: { type: string, format: date-time, nullable: true }
        edit_count: { type: integer }
        is_addendum_of_review_id: { type: string, format: uuid, nullable: true }
...
paths:
...
  /reviews:
    post:
      tags: [review]
      x-required-permission: review.take_seat
      summary: Open a SurgeonReview (acquires reviewer seat per FR-017a)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                analysis_id: { type: string, format: uuid }
      responses:
        '201':
          description: Seat acquired
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SurgeonReview' }
        '409':
          description: Seat held by another user; returns holder display_name
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /reviews/{review_id}/heartbeat:
    parameters:
      - { name: review_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [review]
      summary: Extend seat (called every 60s by client)
      responses:
        '204': { description: Extended }
  /reviews/{review_id}/mask-refine:
    parameters:
      - { name: review_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [review]
      x-required-permission: review.refine_mask
      summary: VISTA3D click-to-refine (FR-015). Local recompute ≤30s.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                segmentation_id: { type: string, format: uuid }
                click_point_voxel: { type: array, items: { type: integer }, minItems: 3, maxItems: 3 }
                click_mode: { type: string, enum: [add, subtract] }
      responses:
        '200':
          description: Local mask updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  segmentation_id: { type: string, format: uuid }
                  new_volume_ml: { type: number }
                  mask_url: { type: string, format: uri }
                  duration_ms: { type: integer }
  /reviews/{review_id}/lesion-prompt:
    parameters:
      - { name: review_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [review]
      x-required-permission: review.reprompt_lesion
      summary: MedSAM-2 one-prompt tumor re-segmentation (FR-016)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                marker_voxel: { type: array, items: { type: integer }, minItems: 3, maxItems: 3 }
      responses:
        '201':
          description: New lesion added to list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lesion' }
  /reviews/{review_id}/classification-override:
    parameters:
      - { name: review_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [review]
      x-required-permission: review.override_classification
      summary: Reviewer overrides AI classification (spec Edge Cases)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                lesion_id: { type: string, format: uuid }
                new_class: { type: string, enum: [hcc, icc, metastasis, fnh, hemangioma, cyst] }
                reason: { type: string }
      responses:
        '200': { description: Override recorded }
  /reviews/{review_id}/flr:
    parameters:
      - { name: review_id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [review]
      summary: Commit a new resection-plane FLR (FR-012/013). Client computes locally; server persists.
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FLRCalculation' }
      responses:
        '201':
          description: Saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FLRCalculation' }
...